requirements.txt
yfinance
google-cloud-storage
apache-airflow
gcsfs